Mailgun - enable custom headers (list-unsubscribe)
The Mailgun Service Handler does not currently check for or honor custom headers that have been set in phpmailer.  This caused the "list-unsubscribe" header from FluentCRM to be lost when sending via Mailgun.

Used and slightly modified code from Elastic Mail handler to keep coding consistent to enable this in the Mailgun Handler.